Admin: Add some additional clarity to some about page links
It's important for links to have enough information in them in order to be clearly actionable. Destination and purpose should be clear. For example, `debug` isn't nearly as useful as `learn how to debug issues`.

This was signed off on by chanthaboune and jorbin.

Props afercia.
See: #46901.


Built from https://develop.svn.wordpress.org/trunk@45274


git-svn-id: http://core.svn.wordpress.org/trunk@45083 1a063a9b-81f0-0310-95a4-ce76da25c4cd