Build/Test Tools: Check if all the required PHP extensions are loaded before running the test suite.
Add the GD extension as a hard requirement.

This improves the reliability of the test suite and ensures that if the test infrastructure changes in the future and a platform requirement such as GD accidentally gets removed, the tests fail with an appropriate error message.

Follow-up to [48592].

Props ayeshrajans, jrf, johnbillion.
Fixes #50640.
Built from https://develop.svn.wordpress.org/trunk@49535


git-svn-id: http://core.svn.wordpress.org/trunk@49273 1a063a9b-81f0-0310-95a4-ce76da25c4cd