Build/Test Tools: Prevent error when workflows run for new branches and tags.
This fixes a JavaScript error in the Slack notifications workflow. This was causing the first workflow run for each workflow in a new branch or tag to be reported as a failure, even if the important parts of the workflow succeed.

When running a workflow for the first time, a default of `failure` is now used for the previous outcome. This will trigger a success notification to be sent to report successful branching/tagging.

Props davidbaumwald.
Fixes #54742.
Built from https://develop.svn.wordpress.org/trunk@53077


git-svn-id: http://core.svn.wordpress.org/trunk@52666 1a063a9b-81f0-0310-95a4-ce76da25c4cd